tools: merge the two finish tools into one `finish`
- handover_ticket folds into finish as an optional `handover` argument
- `handover`/`task` become reserved argument names, optional in the schema
- handover-only registration is gone; role prompts now require the chaining
- python drops the HandoverTicketTool export, finish ships auto-registered
