+++ README.md
@@ -343,7 +343,7 @@ let schema = Schema::parse(json!({
 tickets.ticket(Ticket::new("Write a report.").schema(schema));
 ```
 
-Register a schema per label with `tickets.schema_for_label(label, schema)`: every ticket of that label validates against it (unless built with its own), so work reached by `handover_ticket` is checked like anything submitted directly.
+Register a schema per label with `tickets.schema_for_label(label, schema)`: every ticket of that label validates against it.
 
 ### Compaction
 
@@ -394,8 +394,7 @@ Give agents access to tools. Each tool exposes an action the agent can choose to
 | | `ListDirectoryTool` | List files and directories. |
 | **Shell** | `BashTool` | Run a shell command matching an allowed pattern. |
 | **Web** | `FetchUrlTool` | Fetch a URL and read its body. |
-| **Tickets** | `FinishTicketTool` | Write the result for the current ticket and mark it finished. |
-| | `HandoverTicketTool` | Write the result, mark the ticket finished, and hand follow-up work to another agent. |
+| **Tickets** | `FinishTool` | Write the result for the current ticket and mark it finished, optionally handing follow-up work to another agent. |
 | | `ManageTicketsTool` | Read the ticket queue and create or edit tickets. |
 | | `ReadTicketsTool` | Read the ticket queue. |
 | **Knowledge** | `ManageKnowledgeTool` | Write, read, remove, or list pages in the agent's knowledge store. Registered automatically on every agent. |
