+++ agentdocs/architecture.md
@@ -31,13 +31,15 @@ The invariants that shape how code fits together. Layout says where code lives; 
 
 ## Finishing is a tool call
 
-**Agents finish tickets through one of two finisher tools: `finish_ticket` (terminal) or `handover_ticket` (terminal + spawn child). Both record the result through `TicketSystem::set_result`, which owns the result-validation-and-logging contract, then transition the ticket to `Finished` (`finish_ticket` via the `write_result` helper; `handover_ticket` directly, around the child insert). The loop enforces the rule: a turn that ends without a finisher tool call is rejected and retried.**
+**Agents finish tickets through one tool, `finish`. It records the result through `TicketSystem::set_result`, which owns the result-validation-and-logging contract, then transitions the ticket to `Finished`. An optional `handover` argument additionally spawns a child ticket; its presence is the only discriminator, so there is no second tool and no mode field. The loop enforces the rule: a turn that ends without a `finish` call is rejected and retried.**
 
-- `finish_ticket` writes a `result`, attaches it to the current ticket, and transitions to `Finished`. Use it for terminal work.
-- `handover_ticket` does the same and then inserts a child ticket pinned to a target agent or label with the current ticket recorded as its `parent`. It inserts the child BEFORE finishing the parent, so the child is already `Todo` when the parent leaves the queue: a concurrent `pending_count()` poll can never see an empty queue between them and `finish()` cannot drain the chain early. `TicketFinished` / `TicketFailed` are emitted synchronously from the status transition itself, and an in-flight transition counter holds the drain open until every handler returns, so a `create_ticket_on_result` follow-up is inserted before `finish()` can observe an empty queue. Use it to finish-and-chain in one call; the alternative — `finish_ticket` followed by `manage_tickets_tool::create` — is order-sensitive the other way and leaves the current ticket re-picked when the order is wrong.
-- When a turn ends without a finisher tool call (no `finish_ticket`, no `handover_ticket`, no result attached), the loop pushes a corrective directive and retries. This is the same retry path used for schema-validation failures, bounded by `max_schema_retries`; exhaustion emits `PolicyViolated { MaxSchemaRetries, .. }` and `TicketFailed`.
-- `Status` transitions go through tickets-side helpers; the agent never writes status directly. `Failed` is reserved for system-driven outcomes (schema-retry trip, missing-finisher exhaustion, policy violations).
-- `Ticket::schema(...)` attaches a `Schema` to the ticket; the finisher tool validates the result and the loop applies `max_schema_retries` on mismatch. A schema can also be registered as a per-label default via `TicketSystem::schema_for_label`, stamped onto a schemaless ticket at creation, so a result contract follows its label however the ticket was created (direct, labeled, or a `handover_ticket` child). `handover_ticket` accepts any JSON `result` and validates it against the parent ticket's own schema, like `finish_ticket`; it does not carry a schema for the child, which inherits one only through its label. A schema mismatch in `handover_ticket` aborts before the child is inserted, so neither the parent's finish nor the child happens: the operation stays atomic.
+- Without `handover`, `finish` writes a `result`, attaches it to the current ticket, and transitions to `Finished` via the `write_result` helper. This is terminal work.
+- With `handover`, it does the same and then inserts a child ticket pinned to that agent or label with the current ticket recorded as its `parent`. It inserts the child BEFORE finishing the parent, so the child is already `Todo` when the parent leaves the queue: a concurrent `pending_count()` poll can never see an empty queue between them and `finish()` cannot drain the chain early. `TicketFinished` / `TicketFailed` are emitted synchronously from the status transition itself, and an in-flight transition counter holds the drain open until every handler returns, so a `create_ticket_on_result` follow-up is inserted before `finish()` can observe an empty queue. The alternative — a plain `finish` followed by `manage_tickets_tool::create` — is order-sensitive the other way and leaves the current ticket re-picked when the order is wrong.
+- An agent that must always chain can no longer be forced to by its tool registry, since every `finish` accepts an optional `handover`. Its role prompt carries that requirement instead, which makes those prompt lines load-bearing.
+- When a turn ends without a `finish` call and no result attached, the loop pushes a corrective directive and retries. This is the same retry path used for schema-validation failures, bounded by `max_schema_retries`; exhaustion emits `PolicyViolated { MaxSchemaRetries, .. }` and `TicketFailed`.
+- `Status` transitions go through tickets-side helpers; the agent never writes status directly. `Failed` is reserved for system-driven outcomes (schema-retry trip, missing-`finish` exhaustion, policy violations).
+- `Ticket::schema(...)` attaches a `Schema` to the ticket; `finish` validates the result and the loop applies `max_schema_retries` on mismatch. A schema can also be registered as a per-label default via `TicketSystem::schema_for_label`, stamped onto a schemaless ticket at creation, so a result contract follows its label however the ticket was created (direct, labeled, or a handover child). A handover validates its `result` against the parent ticket's own schema, exactly as a plain finish does; it carries no schema for the child, which inherits one only through its label. A schema mismatch aborts before the child is inserted, so neither the parent's finish nor the child happens: the operation stays atomic.
+- `handover` and `task` are reserved argument names for `finish`. A ticket whose schema is an object has its fields passed as `finish`'s top-level arguments, so such a schema must not declare a `handover` or `task` property: those names are stripped as control keys before the result is recovered.
 - A successful finish appends one NDJSON record `{ticket, result}` to `<dir>/results.jsonl` (configured via `TicketSystem::dir(d)`; defaults to `./.agentwerk`) and attaches the same `result` value to the ticket. The value is surfaced through `Ticket::result()`; `last_result()` returns its serialized form for the most recent `Finished` ticket.
 - The system also appends one JSON line to `<dir>/tickets.jsonl` per lifecycle event (`created`, `started`, `done`, `failed`) and writes the full ticket state to `<dir>/tickets/<key>/ticket.<ts>.json`. The `created` event carries the optional `parent` key when set, giving the log a complete handover audit trail. The log is observational: errors are swallowed. The result payload stays in `results.jsonl`; `tickets.jsonl` carries only the transition.
