+++ crates/agentwerk-py/python/agentwerk/__init__.py
@@ -27,7 +27,6 @@ from ._agentwerk import (
     ListDirectoryTool,
     CodegrepTool,
     FetchUrlTool,
-    HandoverTicketTool,
     ReadTicketsTool,
     ManageTicketsTool,
     BashTool,
@@ -80,7 +79,6 @@ __all__ = [
     "ListDirectoryTool",
     "CodegrepTool",
     "FetchUrlTool",
-    "HandoverTicketTool",
     "ReadTicketsTool",
     "ManageTicketsTool",
     "BashTool",
