+++ crates/agentwerk-py/python/agentwerk/__init__.pyi
@@ -31,7 +31,6 @@ def GlobTool() -> Tool: ...
 def ListDirectoryTool() -> Tool: ...
 def CodegrepTool() -> Tool: ...
 def FetchUrlTool() -> Tool: ...
-def HandoverTicketTool() -> Tool: ...
 def ReadTicketsTool() -> Tool: ...
 def ManageTicketsTool() -> Tool: ...
 def BashTool(
