--- crates/agentwerk/src/tools/tickets/finish_ticket.rs
@@ -1,412 +1,0 @@
-//! Single-purpose tool for finishing a ticket. Validates the agent's
-//! result against the ticket's schema (when set), appends an NDJSON
-//! line to `<dir>/results.jsonl`, attaches the `TicketResult`
-//! to the ticket, and transitions the ticket to `Finished`.
-
-use std::future::Future;
-use std::pin::Pin;
-use std::sync::OnceLock;
-
-use serde_json::Value;
-
-use crate::providers::ProviderResult;
-
-use super::super::tool::{ToolContext, ToolLike, ToolResult};
-use super::super::tool_file::ToolFile;
-use super::{resolve_current_key, write_result};
-
-/// Write a ticket's result and mark it finished.
-///
-/// # Examples
-///
-/// ```
-/// use agentwerk::Agent;
-/// use agentwerk::tools::FinishTicketTool;
-///
-/// Agent::new().tool(FinishTicketTool);
-/// ```
-pub struct FinishTicketTool;
-
-fn tool_file() -> &'static ToolFile {
-    static FILE: OnceLock<ToolFile> = OnceLock::new();
-    FILE.get_or_init(|| ToolFile::parse(include_str!("finish_ticket.tool.md")))
-}
-
-fn description() -> &'static str {
-    static DESC: OnceLock<String> = OnceLock::new();
-    DESC.get_or_init(|| tool_file().render_markdown())
-}
-
-impl ToolLike for FinishTicketTool {
-    fn name(&self) -> &str {
-        &tool_file().name
-    }
-
-    fn description(&self) -> &str {
-        description()
-    }
-
-    fn input_schema(&self) -> Value {
-        tool_file().input_schema.clone()
-    }
-
-    fn is_read_only(&self) -> bool {
-        tool_file().read_only
-    }
-
-    fn call<'a>(
-        &'a self,
-        input: Value,
-        ctx: &'a ToolContext,
-    ) -> Pin<Box<dyn Future<Output = ProviderResult<ToolResult>> + Send + 'a>> {
-        Box::pin(async move {
-            let Some(ticket_system) = ctx.ticket_system_handle().cloned() else {
-                return Ok(ToolResult::error(
-                    "Ticket system unavailable in this context",
-                ));
-            };
-            let key = match resolve_current_key(&ticket_system, ctx) {
-                Ok(k) => k,
-                Err(e) => return Ok(e),
-            };
-            let schema = ticket_system.get_ticket(&key).and_then(|t| t.schema);
-            let agent = ctx.agent_name_str().unwrap_or_default().to_string();
-            let result =
-                super::result_shape::parse_result("finish_ticket", schema.as_ref(), &input);
-            Ok(write_result(&ticket_system, &key, result, &agent))
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-    use std::sync::Arc;
-
-    use super::*;
-    use crate::agents::tickets::{Status, Ticket, TicketSystem};
-    use crate::schemas::Schema;
-
-    fn ctx_with(ticket_system: Arc<TicketSystem>, agent: &str, dir: PathBuf) -> ToolContext {
-        ToolContext::new(dir)
-            .ticket_system(ticket_system)
-            .agent_name(agent.to_string())
-    }
-
-    fn one_ticket(agent: &str) -> (Arc<TicketSystem>, String) {
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        sys.insert(Ticket::new("body").label(agent), "tester".into());
-        let key = sys
-            .claim(|t| t.status == Status::Todo, agent)
-            .expect("claim must succeed");
-        (sys, key)
-    }
-
-    /// Process-lifetime tempdir used as the default `TicketSystem` root
-    /// for tests in this module. Tests that need an isolated workspace
-    /// still call `sys.dir(...)` explicitly to override.
-    fn shared_test_dir() -> &'static std::path::Path {
-        use std::sync::OnceLock;
-        static DIR: OnceLock<crate::test_util::TempDir> = OnceLock::new();
-        DIR.get_or_init(|| crate::test_util::TempDir::new().unwrap())
-            .path()
-    }
-
-    #[tokio::test]
-    async fn writes_string_result_and_marks_finished() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let (sys, key) = one_ticket("alice");
-        sys.dir(dir.path().to_path_buf());
-        let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"result": "the answer"}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::Success(_)));
-        let t = sys.get_ticket(&key).unwrap();
-        assert_eq!(t.status, Status::Finished);
-        assert_eq!(
-            t.result.as_ref().and_then(|v| v.as_str()),
-            Some("the answer")
-        );
-
-        let log = std::fs::read_to_string(dir.path().join("results.jsonl")).unwrap();
-        let line = log.trim_end();
-        let parsed: serde_json::Value = serde_json::from_str(line).unwrap();
-        assert_eq!(parsed["ticket"], key.as_str());
-        assert_eq!(parsed["result"], "the answer");
-    }
-
-    #[tokio::test]
-    async fn accepts_any_value_when_no_schema() {
-        for value in [
-            serde_json::json!(""),
-            serde_json::json!(null),
-            serde_json::json!({}),
-            serde_json::json!([]),
-        ] {
-            let dir = crate::test_util::TempDir::new().unwrap();
-            let (sys, key) = one_ticket("alice");
-            sys.dir(dir.path().to_path_buf());
-            let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-            let outcome = FinishTicketTool
-                .call(serde_json::json!({"result": value}), &ctx)
-                .await
-                .unwrap();
-            assert!(
-                matches!(outcome, ToolResult::Success(_)),
-                "expected success for {value:?}"
-            );
-            let t = sys.get_ticket(&key).unwrap();
-            assert_eq!(t.status, Status::Finished);
-        }
-    }
-
-    #[tokio::test]
-    async fn accepts_structured_value_when_no_schema() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let (sys, key) = one_ticket("alice");
-        sys.dir(dir.path().to_path_buf());
-        let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"result": {"x": 1, "y": [2, 3]}}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::Success(_)));
-        let t = sys.get_ticket(&key).unwrap();
-        assert_eq!(t.status, Status::Finished);
-        assert_eq!(t.result.as_ref().unwrap()["x"], 1);
-
-        // The saved `result` field is a JSON object, not an escaped
-        // string of JSON.
-        let log = std::fs::read_to_string(dir.path().join("results.jsonl")).unwrap();
-        let parsed: serde_json::Value = serde_json::from_str(log.trim_end()).unwrap();
-        assert!(
-            parsed["result"].is_object(),
-            "expected raw object, got {parsed}"
-        );
-        assert_eq!(parsed["result"]["x"], 1);
-    }
-
-    #[tokio::test]
-    async fn validates_against_schema() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        sys.dir(dir.path().to_path_buf());
-        let schema = Schema::parse(serde_json::json!({
-            "type": "object",
-            "properties": {"x": {"type": "string"}},
-            "required": ["x"]
-        }))
-        .unwrap();
-        sys.insert(
-            Ticket::new("hi").schema(schema).label("alice"),
-            "tester".into(),
-        );
-        let key = sys
-            .claim(|t| t.status == Status::Todo, "alice")
-            .expect("claim must succeed");
-        let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-
-        // wrong shape
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"result": {"x": 7}}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::SchemaError(_)));
-        let t = sys.get_ticket(&key).unwrap();
-        assert_eq!(t.status, Status::InProgress);
-
-        // valid shape
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"result": {"x": "ok"}}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::Success(_)));
-        let t = sys.get_ticket(&key).unwrap();
-        assert_eq!(t.status, Status::Finished);
-        assert_eq!(t.result.as_ref().unwrap()["x"], "ok");
-    }
-
-    #[tokio::test]
-    async fn object_schema_takes_flat_arguments_as_the_result() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        sys.dir(dir.path().to_path_buf());
-        let schema = Schema::parse(serde_json::json!({
-            "type": "object",
-            "properties": {"x": {"type": "string"}},
-            "required": ["x"]
-        }))
-        .unwrap();
-        sys.insert(
-            Ticket::new("hi").schema(schema).label("alice"),
-            "tester".into(),
-        );
-        let key = sys
-            .claim(|t| t.status == Status::Todo, "alice")
-            .expect("claim must succeed");
-        let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-
-        // Fields at the top level, no `result` wrapper.
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"x": "ok"}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::Success(_)));
-        let t = sys.get_ticket(&key).unwrap();
-        assert_eq!(t.status, Status::Finished);
-        assert_eq!(t.result.as_ref().unwrap(), &serde_json::json!({"x": "ok"}));
-    }
-
-    #[tokio::test]
-    async fn stores_a_string_encoded_result_as_the_decoded_object() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        sys.dir(dir.path().to_path_buf());
-        let schema = Schema::parse(serde_json::json!({
-            "type": "object",
-            "properties": {"x": {"type": "string"}},
-            "required": ["x"]
-        }))
-        .unwrap();
-        sys.insert(
-            Ticket::new("hi").schema(schema).label("alice"),
-            "tester".into(),
-        );
-        let key = sys
-            .claim(|t| t.status == Status::Todo, "alice")
-            .expect("claim must succeed");
-        let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-
-        // The agent double-encoded the conforming object as a JSON string.
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"result": "{\"x\": \"ok\"}"}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::Success(_)));
-
-        let t = sys.get_ticket(&key).unwrap();
-        assert_eq!(t.status, Status::Finished);
-        // Stored as the decoded object, not the raw string.
-        assert!(t.result.as_ref().unwrap().is_object());
-        assert_eq!(t.result.as_ref().unwrap()["x"], "ok");
-        // The persisted log line carries the object too.
-        let log = std::fs::read_to_string(dir.path().join("results.jsonl")).unwrap();
-        let parsed: serde_json::Value =
-            serde_json::from_str(log.lines().next_back().unwrap()).unwrap();
-        assert!(parsed["result"].is_object());
-    }
-
-    #[tokio::test]
-    async fn errors_when_no_current_ticket() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        let ctx = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-        let outcome = FinishTicketTool
-            .call(serde_json::json!({"result": "x"}), &ctx)
-            .await
-            .unwrap();
-        assert!(matches!(outcome, ToolResult::Error(_)));
-    }
-
-    #[tokio::test]
-    async fn appends_one_line_per_completed_ticket() {
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        sys.dir(dir.path().to_path_buf());
-
-        sys.insert(Ticket::new("a").label("alice"), "tester".into());
-        let key1 = sys
-            .claim(|t| t.key == "TICKET-1", "alice")
-            .expect("claim must succeed");
-        let ctx_alice = ctx_with(Arc::clone(&sys), "alice", dir.path().to_path_buf());
-        FinishTicketTool
-            .call(serde_json::json!({"result": "from alice"}), &ctx_alice)
-            .await
-            .unwrap();
-
-        sys.insert(Ticket::new("b").label("bob"), "tester".into());
-        let key2 = sys
-            .claim(|t| t.key == "TICKET-2", "bob")
-            .expect("claim must succeed");
-        let ctx_bob = ctx_with(Arc::clone(&sys), "bob", dir.path().to_path_buf());
-        FinishTicketTool
-            .call(serde_json::json!({"result": "from bob"}), &ctx_bob)
-            .await
-            .unwrap();
-
-        let log = std::fs::read_to_string(dir.path().join("results.jsonl")).unwrap();
-        let lines: Vec<&str> = log.lines().collect();
-        assert_eq!(lines.len(), 2);
-        let first: serde_json::Value = serde_json::from_str(lines[0]).unwrap();
-        let second: serde_json::Value = serde_json::from_str(lines[1]).unwrap();
-        assert_eq!(first["ticket"], key1.as_str());
-        assert_eq!(first["result"], "from alice");
-        assert_eq!(second["ticket"], key2.as_str());
-        assert_eq!(second["result"], "from bob");
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn concurrent_writes_produce_one_intact_line_per_ticket() {
-        const N: usize = 32;
-        let dir = crate::test_util::TempDir::new().unwrap();
-        let sys = TicketSystem::new();
-        sys.dir(shared_test_dir().to_path_buf());
-        sys.dir(dir.path().to_path_buf());
-
-        let mut expected = Vec::with_capacity(N);
-        for i in 0..N {
-            let agent = format!("agent_{i}");
-            sys.insert(
-                Ticket::new(format!("body_{i}")).label(&agent),
-                "tester".into(),
-            );
-            let key = sys
-                .claim(
-                    |t| t.status == Status::Todo && t.labels.iter().any(|l| l == &agent),
-                    &agent,
-                )
-                .expect("claim must succeed");
-            expected.push((agent, key));
-        }
-
-        let mut handles = Vec::with_capacity(N);
-        for (i, (agent, _)) in expected.iter().enumerate() {
-            let sys = Arc::clone(&sys);
-            let dir_path = dir.path().to_path_buf();
-            let agent = agent.clone();
-            handles.push(tokio::spawn(async move {
-                let ctx = ctx_with(sys, &agent, dir_path);
-                FinishTicketTool
-                    .call(serde_json::json!({"result": format!("payload_{i}")}), &ctx)
-                    .await
-                    .unwrap()
-            }));
-        }
-        for h in handles {
-            assert!(matches!(h.await.unwrap(), ToolResult::Success(_)));
-        }
-
-        let log = std::fs::read_to_string(dir.path().join("results.jsonl")).unwrap();
-        let lines: Vec<&str> = log.lines().collect();
-        assert_eq!(lines.len(), N, "expected {N} lines, got {}", lines.len());
-
-        let mut seen_tickets = std::collections::HashSet::new();
-        for line in &lines {
-            let parsed: serde_json::Value =
-                serde_json::from_str(line).unwrap_or_else(|e| panic!("corrupt line {line:?}: {e}"));
-            let ticket = parsed["ticket"].as_str().unwrap().to_string();
-            assert!(seen_tickets.insert(ticket), "duplicate ticket in log");
-        }
-        let expected_keys: std::collections::HashSet<String> =
-            expected.iter().map(|(_, k)| k.clone()).collect();
-        assert_eq!(seen_tickets, expected_keys);
-    }
-}
