+++ crates/agentwerk/tests/integration/glob_finds_nested_files.rs
@@ -94,7 +94,7 @@ async fn finds_every_lib_rs_in_nested_tree() -> std::result::Result<(), Box<dyn 
                 "Investigate the working directory and answer the user's question. \
                  Use the available tools — pick whichever one fits the question. \
                  When you have the answer, settle the ticket via \
-                 `finish_ticket`.",
+                 `finish`.",
             )
             .tool(GlobTool)
             .tool(GrepTool)
