+++ crates/agentwerk/tests/integration/grep_finds_code_pattern.rs
@@ -103,7 +103,7 @@ async fn finds_code_pattern_with_special_chars(
                 "Investigate the working directory and answer the user's question. \
                  Use the available tools — pick whichever one fits the question. \
                  When you have the answer, settle the ticket via \
-                 `finish_ticket`.",
+                 `finish`.",
             )
             .tool(GrepTool)
             .tool(GlobTool)
