+++ crates/agentwerk/tests/integration/list_directory_enumerates_entries.rs
@@ -51,7 +51,7 @@ async fn separates_files_and_directories() -> std::result::Result<(), Box<dyn st
         .role(
             "Step 1: call `list_directory_tool` with `path: \".\"` to see the \
              working directory's top level. \
-             Step 2: immediately call `finish_ticket` with `files` and \
+             Step 2: immediately call `finish` with `files` and \
              `directories` as its top-level arguments in exactly this shape: \
              {\"files\": [\"<basename>\", ...], \"directories\": [\"<basename>\", ...]}. \
              Never prose, never a bullet list, never a sentence. Do not output \
