+++ crates/agentwerk/tests/integration/read_tickets_finds_seeded.rs
@@ -34,7 +34,7 @@ async fn reports_a_secret_from_another_ticket(
             .model(&model)
             .role(
                 "Answer the user by inspecting the ticket queue with the \
-                 available ticket tool, then call `finish_ticket` with the \
+                 available ticket tool, then call `finish` with the \
                  answer. Do not guess: read the relevant ticket and quote it.",
             )
             .tool(ReadTicketsTool)
