+++ crates/agentwerk/tests/integration/traces_call_path_across_files.rs
@@ -89,7 +89,7 @@ async fn traces_three_hop_call_path() -> std::result::Result<(), Box<dyn std::er
              available read-only tools as you see fit. Multiple functions in \
              this codebase share names: rely on the source of each function, \
              not just its name, to decide what calls what. When you have the \
-             full chain, finish the ticket via `finish_ticket` with the \
+             full chain, finish the ticket via `finish` with the \
              result shaped to match the schema.",
         )
         .tool(GrepTool)
