+++ crates/agentwerk/tests/integration/write_file_creates_file.rs
@@ -28,7 +28,7 @@ async fn creates_file_with_token() -> std::result::Result<(), Box<dyn std::error
         .role(
             "Step 1: call `write_file_tool` to create exactly the file the user \
              asks for, with exactly the content they specify (and nothing else). \
-             Step 2: immediately call `finish_ticket` to settle the \
+             Step 2: immediately call `finish` to settle the \
              ticket. Do not write any prose — your only output must be tool \
              calls.",
         )
