+++ crates/use-cases/src/divide_and_conquer/prompts/agent.role.md
@@ -4,5 +4,5 @@ If the tool fails or returns something other than a single integer, say so rathe
 
 - Each task body gives the bounds `lo`, `hi`, and a partition index `idx`; substitute the numeric bounds in every directive below.
 - MUST call `python` with `{"code": "print(sum(k*k for k in range(LO, HI + 1)))"}`, substituting the bounds from the task.
-- Finish the ticket by calling `finish_ticket` with `idx` and `partial_sum` as its top-level arguments: `finish_ticket({"idx": IDX, "partial_sum": N})`, copying `idx` verbatim from the task and using the integer the tool printed for `N`.
-- NEVER add prose, code fences, or commentary outside the `finish_ticket` call.
+- Finish the ticket by calling `finish` with `idx` and `partial_sum` as its top-level arguments: `finish({"idx": IDX, "partial_sum": N})`, copying `idx` verbatim from the task and using the integer the tool printed for `N`.
+- NEVER add prose, code fences, or commentary outside the `finish` call.
