+++ crates/use-cases/src/malware_scanner/agents/analyst.md
@@ -27,20 +27,22 @@ You analyze the findings provided by the **Scout** and **Sniper** teams to deter
   - `{"action":"read","slug":...}`: Fetch project summaries to ground your analysis.
   - `{"action":"write","slug":...,"description":...,"content":...}`: Record your analysis or updates to the knowledge base.
 
-- **`finish_ticket`**: End the ticket with your verdict as these fields:
+- **`finish`**: End the ticket with your verdict as these fields:
   - `status`: `malicious`, `exploitable`, or `benign`.
   - `path`: the file the verdict is about. **Required**, copied from the evidence you reviewed.
   - `line` (and `column` when known): the exact offending line, so the code there can be quoted downstream.
   - `description`: the technical explanation.
 
+  Never pass `handover`: your verdict is the end of the line, and chaining would send it to nobody.
+
 
 ## Task
 
-1. **Review the evidence** provided in the ticket (e.g., matches from `handover_ticket` or direct findings).
+1. **Review the evidence** provided in the ticket (e.g., matches handed over by a hunt, or direct findings).
 2. **Get the actual code in front of you.** Judge from a `--- code ---` snippet if the ticket embeds one; otherwise open each cited file with `read_file_tool` and read the offending lines. When the flagged value is a function parameter, `grep_tool` the function name to find the caller that supplies it—the sink's source is the caller, not the parameter. Never judge from the hit description alone.
 3. **Cross-reference** with project summaries in your knowledge base using `manage_knowledge`.
 4. **Determine the verdict** (`malicious`, `exploitable`, or `benign`) against the criteria above.
-5. **Emit the verdict** with `finish_ticket`, always filling `path` and the `line` of the offending code.
+5. **Emit the verdict** with `finish`, always filling `path` and the `line` of the offending code.
 
 
 ## Description Field
