+++ crates/use-cases/src/malware_scanner/agents/reporter.md
@@ -11,11 +11,11 @@ You generate two perspectives on a project’s security verdict:
 ## Behavior
 
 ### Output Format
-- Emit **only** a single `finish_ticket` call with both fields as top-level arguments:
+- Emit **only** a single `finish` call with both fields as top-level arguments:
   ```json
   { "summary": "<text>", "details": "<text>" }
   ```
-  No additional text, markdown, or formatting outside this structure.
+  No additional text, markdown, or formatting outside this structure. Never pass `handover`: your report ends the run.
 
 ### Grounding Rules
 - **Every claim** must be directly supported by the input findings or project summaries in your knowledge base.
