+++ crates/use-cases/src/malware_scanner/agents/scout.md
@@ -30,13 +30,12 @@ A fresh ticket will arrive for the next part, so focus only on the assigned segm
   - `{"action":"read","slug":...}`: Open a page (e.g., the file map).
   - `{"action":"write","slug":...,"description":<≤80 chars>,"content":...}`: Write or update a knowledge page. All fields (`slug`, `description`, `content`) are required and non-empty.
   - Correct a page in place under its slug. **Never delete a page.**
-- **`handover_ticket`**: Flags your lead and ends the ticket. **Every call requires both fields:**
-  - `to`: The target (e.g., `sniper_hunt` or `security_analysis`).
+- **`finish`**: Ends the ticket. When you have a lead, pass **both fields**:
+  - `handover`: The target (e.g., `sniper_hunt` or `security_analysis`).
   - `result`: Your bundle (see **Task** section for format). It becomes the follow-up ticket's body.
-  **Omitting either field will fail the call.**
-- **`finish_ticket`**: Use **only if you found nothing to flag**. Provide a one-line `summary`.
+  **Omitting either field with a lead in hand will lose it.** Only if you found nothing to flag, call `finish` with a one-line `result` and no `handover`.
 
-**Note:** These five tools are your only options. Any other tool name will fail and waste your turn.
+**Note:** These four tools are your only options. Any other tool name will fail and waste your turn.
 
 
 ## Task
@@ -61,15 +60,15 @@ You have a limited time budget (`{scout_time_budget}`). Follow this sequence **o
      - If the instruction specifies a file type, extension, folder, or concern, **use that as your lead**.
      - Otherwise, flag the **single most notable technology or pattern** discovered.
      ```json
-     handover_ticket({
-       "to": "sniper_hunt",
+     finish({
+       "handover": "sniper_hunt",
        "result": "technology: <name>\nobserved: <path> — <construct>\nhypothesis: <why worth hunting>"
      })
      ```
    - **Rare case (direct line for judgment):**
      ```json
-     handover_ticket({
-       "to": "security_analysis",
+     finish({
+       "handover": "security_analysis",
        "result": "source: <path>\nline: <N>\nexcerpt: <exact text>\nreason: <2-5 sentences>"
      })
      ```
