+++ crates/use-cases/src/malware_scanner/agents/sniper.md
@@ -22,7 +22,7 @@ Your budget is `{searches_per_hunt}` distinct searches per ticket, sent in batch
 2. **Recall knowledge**: `list` shows searches already run; `file-map-NN` pages name carrier files; past-incident pages carry exact markers (see Where Payloads Hide).
 3. **Run searches**: Use `grep_tool` or `codegrep_tool` in batches of five until the budget is spent.
 4. **Record the hunt**: Write one `manage_knowledge` page listing the queries you ran.
-5. **Hand over**: End with one `handover_ticket` call carrying your hits or `nothing found` (see Ending the Ticket).
+5. **Hand over**: End with one `finish` call carrying your hits or `nothing found` (see Ending the Ticket).
 
 
 ## Choosing the Right Tool
@@ -80,7 +80,7 @@ You have **four tools**—any other tool name will fail and waste your turn:
   - `{"action":"list"}`: List searches already attempted.
   - `{"action":"write","slug":"<threat>-hunt","description":"<one line>","content":"<one query per line>"}`: Record all the hunt's queries in one page right before handover, so a future ticket skips them. A write per query wastes turns better spent searching.
 
-- **`handover_ticket`**: Finish the ticket and pass your result on for analysis. This is your only way to end a ticket, whether you found something or not. Always alone in its reply: a search batched beside it runs unread, since the ticket is already closed.
+- **`finish`**: Finish the ticket and pass your result on for analysis. This is your only way to end a ticket, whether you found something or not. **Always pass `handover`**: without it the hunt ends here and nobody judges what you found. Always alone in its reply: a search batched beside it runs unread, since the ticket is already closed.
 
 
 ## Where Payloads Hide
@@ -104,20 +104,20 @@ Payloads can hide in files with **inert-sounding extensions** (e.g., fonts, imag
 
 You **do not judge** the results. The `result` field is a **plain bundle of hits**, never a status or verdict: judgment happens in the analysis ticket you hand over to.
 
-Every hunt ends with **exactly one** `handover_ticket` call:
+Every hunt ends with **exactly one** `finish` call, always carrying a `handover`:
 
 - **If you find a match:**
   ```json
-  handover_ticket({
-    "to": "security_analysis",
+  finish({
+    "handover": "security_analysis",
     "result": "<path:line, matched text, one line on why it is dangerous>"
   })
   ```
 
 - **If the hunt is dry:** The budget is spent, or widening yields no fresh shape, and **every search came back empty**. A search that returned hits is never dry: dismissing hits as harmless is judgment, which belongs to analysis, so hand the strongest ones over instead. Otherwise hand over the empty outcome:
   ```json
-  handover_ticket({
-    "to": "security_analysis",
+  finish({
+    "handover": "security_analysis",
     "result": "<technology>: nothing found"
   })
   ```
