+++ crates/use-cases/src/malware_scanner/main.rs
@@ -20,9 +20,7 @@ use std::time::{Duration, Instant};
 
 use agentwerk::event::{Event, EventKind, PolicyKind};
 use agentwerk::providers::{model_from_env, provider_from_env, Provider};
-use agentwerk::tools::{
-    CodegrepTool, GlobTool, GrepTool, HandoverTicketTool, ListDirectoryTool, ReadFileTool,
-};
+use agentwerk::tools::{CodegrepTool, GlobTool, GrepTool, ListDirectoryTool, ReadFileTool};
 use agentwerk::{Agent, Knowledge, Ticket, TicketSystem};
 use serde_json::{json, Value};
 
@@ -40,12 +38,12 @@ const SCOUT_AGENT: &str = include_str!("agents/scout.md");
 const REPORTER_AGENT: &str = include_str!("agents/reporter.md");
 const REPORTER_LABEL: &str = "reporter";
 
-/// Shared `finish_ticket` calling convention, substituted into every agent
+/// Shared `finish` calling convention, substituted into every agent
 /// whose ticket carries a `Schema`. One source of truth instead of each
 /// agent's `## Output` section hand-writing its own wording, since a model
 /// that leaves a field JSON-encoded as a string (rather than emitting it as
 /// a native array or object) fails schema validation.
-const OUTPUT_CONTRACT: &str = "Call `finish_ticket` exactly once with the result fields as its top-level arguments. Emit each field as its native JSON type, never as a JSON-encoded string: an array field is a JSON array, not a string containing array syntax; a text field is plain text, not a string containing escaped JSON.";
+const OUTPUT_CONTRACT: &str = "Call `finish` exactly once with the result fields as its top-level arguments. Emit each field as its native JSON type, never as a JSON-encoded string: an array field is a JSON array, not a string containing array syntax; a text field is plain text, not a string containing escaped JSON.";
 
 /// Distinct searches a Sniper aims for per hunt before the pool refills it.
 /// Injected into the prompt so the target lives in one place.
@@ -251,8 +249,8 @@ async fn main() {
     // Idle from the start; tickets only arrive once a Scout flags a threat.
     for i in 0..concurrency {
         tickets.agent(
-            // No finish_ticket: every hunt ends in a handover.
-            Agent::empty()
+            // Every hunt ends in a handover; sniper.md is what enforces it.
+            Agent::new()
                 .name(format!("Sniper {}", i + 1))
                 .provider(Arc::clone(&provider))
                 .model(&model)
@@ -264,7 +262,6 @@ async fn main() {
                 .knowledge(&sniper_knowledge)
                 .tool(CodegrepTool)
                 .tool(GrepTool)
-                .tool(HandoverTicketTool)
                 .build(),
         );
     }
@@ -286,7 +283,6 @@ async fn main() {
                 .tool(ReadFileTool)
                 .tool(ListDirectoryTool)
                 .tool(GlobTool)
-                .tool(HandoverTicketTool)
                 .build(),
         );
     }
@@ -649,7 +645,7 @@ mod tests {
                 Ok(ModelResponse {
                     content: vec![ContentBlock::ToolUse {
                         id: "call-1".into(),
-                        name: "finish_ticket".into(),
+                        name: "finish".into(),
                         input: json!({ "result": result }),
                     }],
                     status: ResponseStatus::ToolUse,
