+++ crates/use-cases/src/malware_scanner/report.rs
@@ -183,7 +183,7 @@ pub(crate) fn analyst_result_schema_json() -> Value {
 }
 
 /// Result schema for an Analyst ticket: the verdict object the
-/// analyst must emit. Attaching it makes `finish_ticket` reject a
+/// analyst must emit. Attaching it makes `finish` reject a
 /// stringified or shapeless result at finish time, forcing a retry,
 /// instead of letting it through to be counted as `unparsed` here.
 pub(crate) fn analyst_result_schema() -> Schema {
@@ -192,7 +192,7 @@ pub(crate) fn analyst_result_schema() -> Schema {
 
 /// Result schema for the Reporter ticket: an object carrying the lay-reader
 /// `summary` and the three-paragraph technical `details`. Attaching it makes
-/// `finish_ticket` reject a malformed or partial result and force a retry,
+/// `finish` reject a malformed or partial result and force a retry,
 /// the same guard the analyst verdict relies on. The length floors and caps
 /// sit outside the targets the prompt states, so a compliant result is never
 /// rejected while a skimped or runaway one is.
@@ -413,7 +413,7 @@ pub(crate) fn log_event(event: &Event, tickets: &TicketSystem) {
             eprintln!("{c}[{agent}]{r} {verb}");
         }
         // The completion line reads the schema-validated stored result, not the
-        // raw finish_ticket input the analyst may have JSON-encoded as a string.
+        // raw finish input the analyst may have JSON-encoded as a string.
         EventKind::TicketFinished => {
             if let Some(result) = tickets.get_ticket(&event.ticket_key).and_then(|t| t.result) {
                 eprintln!("{c}[{agent}]{r} {}", finish_summary(&result));
@@ -427,8 +427,9 @@ pub(crate) fn log_event(event: &Event, tickets: &TicketSystem) {
         EventKind::ToolCallStarted {
             tool_name, input, ..
         } => {
-            // finish_ticket is reported at TicketFinished from the stored result.
-            if tool_name != "finish_ticket" {
+            // A plain finish is reported at TicketFinished from the stored
+            // result; a handover still prints, since it names who picks it up.
+            if tool_name != "finish" || handover_label(input).is_some() {
                 eprintln!("{c}[{agent}]{r} {}", tool_call_summary(tool_name, input));
             }
         }
@@ -468,6 +469,11 @@ pub(crate) fn log_event(event: &Event, tickets: &TicketSystem) {
         }
         _ => {}
     }
+}
+
+/// The label a `finish` call hands its child ticket to, when it chains one.
+fn handover_label(input: &Value) -> Option<&str> {
+    input.get("handover").and_then(|v| v.as_str())
 }
 
 fn tool_call_summary(tool_name: &str, input: &Value) -> String {
@@ -572,8 +578,8 @@ fn tool_call_summary(tool_name: &str, input: &Value) -> String {
                 other => other.into(),
             }
         }
-        "handover_ticket" => {
-            let to = input.get("to").and_then(|v| v.as_str()).unwrap_or("?");
+        "finish" => {
+            let to = handover_label(input).unwrap_or("?");
             let result = input
                 .get("result")
                 .and_then(|v| v.as_str())
